update contact page layout
